# backend/docs/swagger.yaml
basePath: /
definitions:
...
  kategoriModel.Kategori:
    properties:
      deskripsi:
        type: string
      id:
        type: string
      nama:
        type: string
      ref_kategori:
        $ref: "#/definitions/kategoriModel.RefKategori"
    type: object
...
  kategoriModel.KategoriReference:
    properties:
      deskripsi:
        type: string
      id:
        type: string
      nama:
        type: string
      ref_kategori_id:
        type: string
    type: object
  kategoriModel.KategoriShow:
    properties:
      deskripsi:
        type: string
      id:
        type: string
      nama:
        type: string
      ref_kategori:
        $ref: "#/definitions/kategoriModel.RefKategoriShow"
    type: object
  kategoriModel.RefKategori:
    properties:
      id:
        type: string
      nama:
        type: string
    type: object
...
  kategoriModel.RefKategoriShow:
    properties:
      id:
        type: string
      nama:
        type: string
    type: object
  kategoriModel.Reference:
    properties:
      id:
        type: string
      nama:
        type: string
      values:
        items:
          $ref: "#/definitions/kategoriModel.KategoriReference"
        type: array
    type: object
...
  model.ReferenceResult:
    properties:
      data:
        items:
          $ref: "#/definitions/kategoriModel.Reference"
        type: array
      last:
        type: string
    type: object
  model.ResponseMessage:
    properties:
      message:
        type: string
      reference:
        $ref: "#/definitions/model.ReferenceResult"
      status:
        type: integer
      success:
        type: boolean
    type: object
  orderModel.Order:
    properties:
      id:
        type: string
      invoice:
        type: string
      nama_customer:
        type: string
      status:
        $ref: "#/definitions/orderModel.status"
      total:
        type: integer
      user:
        $ref: "#/definitions/userModel.User"
      user_id:
        type: string
      waktu:
        type: string
    type: object
  orderModel.OrderAdd:
    properties:
      jumlah:
        type: integer
      katalog_id:
        type: string
      subtotal:
        type: integer
    type: object
  orderModel.OrderItemResult:
    properties:
      created_at:
        type: string
      id:
        type: string
      invoice:
        type: string
      nama_customer:
        type: string
      status:
        type: string
      total:
        type: integer
    type: object
  orderModel.OrderListShow:
    properties:
      id:
        type: string
      jumlah:
        type: integer
      order_id:
        type: string
      produk:
        $ref: "#/definitions/produkModel.ProdukShow"
      total_harga:
        type: integer
    type: object
  orderModel.OrderSetStatus:
    properties:
      status:
        type: string
    type: object
  orderModel.status:
    enum:
      - batal
      - selesai
      - proses
      - belum bayar
      - sudah bayar
    type: string
    x-enum-varnames:
      - batal
      - selesai
      - proses
      - belum
      - sudah
  produkModel.Produk:
    properties:
      detail:
        type: string
      foto:
        type: string
      harga:
        type: integer
      id:
        type: string
      kategori:
        $ref: "#/definitions/kategoriModel.Kategori"
      kategori_id:
        type: string
      nama:
        type: string
    type: object
  produkModel.ProdukShow:
    properties:
      detail:
        type: string
      foto:
        type: string
      harga:
        type: integer
      id:
        type: string
      kategori:
        $ref: "#/definitions/kategoriModel.KategoriShow"
      nama:
        type: string
    type: object
...
  userModel.User:
    properties:
      email:
        type: string
      foto:
        type: string
      id:
        type: string
      kategori:
        $ref: "#/definitions/kategoriModel.Kategori"
      nama:
        type: string
      no_telp:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: dokumentasi API Kavi Kasir
  title: KAVI Kasir API
  version: "1.0"
paths:
...
  /order:
    get:
      consumes:
        - application/json
      description: Menampilkan order list
      parameters:
        - description: page
          in: path
          name: page
          type: integer
        - description: size
          in: path
          name: size
          type: integer
        - description: keyword
          in: path
          name: keyword
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/orderModel.Order"
            type: array
      summary: Menampilkan order list
      tags:
        - order
    post:
      consumes:
        - application/json
      description: Menambahkan order
      parameters:
        - description: Field order
          in: body
          name: body
          required: true
          schema:
            items:
              $ref: "#/definitions/orderModel.OrderAdd"
            type: array
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/model.ResponseMessage"
      summary: Menambahkan order
      tags:
        - order
  /order/{order_id}:
    delete:
      consumes:
        - application/json
      description: hapus order
      parameters:
        - description: order_id
          in: path
          name: order_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/produkModel.Produk"
      summary: Hapus order
      tags:
        - order
    get:
      consumes:
        - application/json
      description: Order
      parameters:
        - description: order_id
          in: path
          name: order_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/orderModel.OrderListShow"
            type: array
      summary: Mendapatkan list order berdasarkan id
      tags:
        - order
    patch:
      consumes:
        - application/json
      description: update status order
      parameters:
        - description: order_id
          in: path
          name: order_id
          required: true
          type: string
        - description: Field order
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/orderModel.OrderSetStatus"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/orderModel.OrderItemResult"
      summary: Update status order
      tags:
        - order
...
swagger: "2.0"